// lib/constants2.d.ts
export declare const MonitConstants: {
    USE_WOTB6: boolean;
    DEFAULT_LANGUAGE: string;
    MIN_WILLMEMBERS_UPDATE_FREQ: number;
    MIN_MEMBERS_UPDATE_FREQ: number;
    STEP_COUNT_MIN: number;
    STEP_COUNT_MAX: number;
    MIN_CACHE_UPDATE_FREQ: number;
    MIN_WOT_QUALITY_CACHE_UPDATE_FREQ: number;
    QUALITY_CACHE_ACTION: {
        GET_QUALITY: number;
        GET_MEANS: number;
        INIT: number;
        GET_SENTRIES_COUNT: number;
        GET_D_SEN: number;
    };
    MEMBERS_VIEW: {
        DEFAULT_MEMBERS_RANDOM_NUMBER: number;
        MEMBERS_DISPLAY_MAX: number;
    };
};